% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.R
\name{make_table3}
\alias{make_table3}
\title{Generate Table 3: Adverse events}
\usage{
make_table3(data)
}
\arguments{
\item{data}{Data frame from simulate_impower133()}
}
\value{
A gt table object
}
\description{
Create a publication-ready adverse events table summarizing
treatment-related adverse events by grade in the as-treated
population, using the gt package.
}
