% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtriang}
\alias{rtriang}
\title{Triangular distribution random number generator}
\usage{
rtriang(n, min_val, mode_val, max_val)
}
\arguments{
\item{n}{Sample size}

\item{min_val}{Minimum value}

\item{mode_val}{Mode (peak location)}

\item{max_val}{Maximum value}
}
\value{
A numeric vector of length n
}
\description{
Generate random numbers following a triangular distribution,
useful for simulating continuous variables whose median is not
at the center of the range.
}
\examples{
rtriang(100, 28, 64, 90)
}
